proto(job): introduce separate (optional) Linking Phase
In linking phase, all SMC peers are connected with each other (mostly O(n^2)).
It makes sense to execute this part separately to detect connection issues
early. Otherwise, this enables more precise measurements so the session phase
only contains the SMC logic and not the initial establishment of communication
channels.