feat(orchestration): more roboust pipeline with some retries
Useful with peers that slowly recreate new communication channels to the GW.
Consequently, to the system, it seems that these peers are fully occupied and cannot
do any work. By retrying with some backoff, the GW can integrate them in a setup.
This is especially of interest if min/max number of peers are defined with debug option set.